backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-auth==2.40.1
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
PyPDF2==3.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
